requirements.txt
streamlit
groq
python-dotenv

langchain
langchain-core
langchain-community
langchain-text-splitters

faiss-cpu
sentence-transformers
pypdf

duckduckgo-search
requests
pandas
numpy